Refactor recipe field name and type

+++ src/main/java/com/what2cook/what2cook/domain/recipe/RecipeRepository.java
@@ -2,5 +2,5 @@ package com.what2cook.what2cook.domain.recipe;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecipeRepository extends JpaRepository<Recipe, Long> {
+public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
 }
